test: add scene context smoke test
